refactor(trace): share the range cube-source helper
