+++ flexviz/trace/box.py
@@ -13,7 +13,7 @@ from typing import Any, Dict
 
 import polars as pl
 
-from ..cube import FreeAxisSpec, temporal_unit
+from ..cube import FreeAxisSpec
 from ..LF import AggregationSpec, GroupedAggregationSpec
 from ..spec import TraceSpec
 from .base import (
@@ -21,41 +21,11 @@ from .base import (
     GroupedChildResult,
     TraceResult,
     _child_uid_for_group,
-    _dtype_for_col,
     _group_value_key,
     _group_values_from_frame,
+    _range_cube_source_spec,
     _to_col_tuple,
 )
-
-
-def _range_cube_source_spec(
-    column: str,
-    axis_range: tuple[float, float] | None,
-    schema: pl.Schema | None,
-) -> FreeAxisSpec | None:
-    """Shared 1-D range cube-source descriptor (hist-shaped; cube plan step 8).
-
-    Continuous/temporal kind from the schema dtype; temporal axes carry their
-    physical ``unit`` (contract G) and unsupported temporal dtypes
-    (``Datetime("ns")``, ``Time``) gate to ``None``, as do non-numeric dtypes.
-    Without a schema the kind defaults to ``"continuous"``. ``domain`` is the
-    viewport range verbatim (``None`` = unzoomed; engine-resolved).
-
-    A twin lives in ``line.py`` — keep the two in sync (a shared home in
-    ``base.py`` is deliberately out of this step's file scope).
-    """
-    dtype = _dtype_for_col(schema, column)
-    if dtype is not None:
-        if dtype.is_temporal():
-            unit = temporal_unit(dtype)
-            if unit is None:
-                return None
-            return FreeAxisSpec(
-                column=column, kind="temporal", p=2048, domain=axis_range, unit=unit
-            )
-        if not dtype.is_numeric():
-            return None
-    return FreeAxisSpec(column=column, kind="continuous", p=2048, domain=axis_range)
 
 
 class BoxPlot(FlexTrace):
